+++ java-client/src/test/java11/org/opensearch/client/opensearch/integTest/httpclient5/PingAndInfoIT.java
@@ -8,6 +8,12 @@
 
 package org.opensearch.client.opensearch.integTest.httpclient5;
 
+import org.opensearch.client.opensearch._types.OpenSearchVersionInfo;
 import org.opensearch.client.opensearch.integTest.AbstractPingAndInfoIT;
 
-public class PingAndInfoIT extends AbstractPingAndInfoIT implements HttpClient5TransportSupport {}
+public class PingAndInfoIT extends AbstractPingAndInfoIT implements HttpClient5TransportSupport {
+    @Override
+    protected String expectedHttpProtocol(OpenSearchVersionInfo version) {
+        return isHttps() && version.number().startsWith("3") ? "HTTP/2.0" : "HTTP/1.1";
+    }
+}